Update README.md
The MLOps project is a comprehensive solution for deploying machine learning models efficiently. It incorporates essential components such as preprocessing, a user-friendly React frontend, a robust Django backend, and Docker containerization. By following the straightforward setup instructions, users can quickly deploy the project and start using its features. Contributions to the project are encouraged, and the MIT License allows for flexibility in usage and modification.